Move scripture typography styles to new home
